refactor: ♻️ move files to folder engine to package root
