+++ packages/engine/src/handlers/JSONHandler.ts
@@ -1,8 +1,8 @@
-import JSONConfig from "../../classes/JSONConfig"
-import EngineError from "../../classes/EngineError"
-import { IExperienceJSONSchema } from "../../types"
-import { IEngineResolvers, IFireboltSession } from "../../interfaces/IEngine"
-import validateJSON from "../../helpers/validateJSON"
+import JSONConfig from "../classes/JSONConfig"
+import EngineError from "../classes/EngineError"
+import { IExperienceJSONSchema } from "../types"
+import { IEngineResolvers, IFireboltSession } from "../interfaces/IEngine"
+import validateJSON from "../helpers/validateJSON"
 
 class JSONHandler {
   private preDefinedJSONConfig?: IExperienceJSONSchema
