+++ packages/engine/src/handlers/SessionHandler.ts
@@ -5,9 +5,9 @@ import {
   IExperienceState,
   IFireboltSession,
   IStepSession,
-} from "../../interfaces/IEngine"
-import JSONConfig from "../../classes/JSONConfig"
-import EngineError from "../../classes/EngineError"
+} from "../interfaces/IEngine"
+import JSONConfig from "../classes/JSONConfig"
+import EngineError from "../classes/EngineError"
 
 class SessionHandler {
   private resolvers: IEngineResolvers
@@ -107,9 +107,15 @@ class SessionHandler {
   }
 
   // atualiza o last completed state
-  async addCompletedStep(stepSlug: string, {fields, processedData}: IStepSession, shouldSaveProcessedData: boolean = false) {
+  async addCompletedStep(
+    stepSlug: string,
+    { fields, processedData }: IStepSession,
+    shouldSaveProcessedData: boolean = false
+  ) {
     const currentState = this.current?.experienceState
-    const definitiveStepSession = shouldSaveProcessedData ? {fields, processedData}: {fields}
+    const definitiveStepSession = shouldSaveProcessedData
+      ? { fields, processedData }
+      : { fields }
     const newStepsCompleted = {
       ...this.current?.steps,
       [stepSlug]: definitiveStepSession,
@@ -121,7 +127,6 @@ class SessionHandler {
         lastCompletedStepSlug: stepSlug,
       },
       steps: newStepsCompleted,
-      
     }
 
     await this.updateSession(newSession)
